Reduce the number of compilation units by adding source files FFlBeams, FFlShells, FFlSolids and FFlLoads containing elements of similar kind
