+++ FFlLib/FFlFEParts/FFlAllFEParts.H
@@ -16,20 +16,10 @@
 #include "FFlCMASS.H"
 #include "FFlSPRING.H"
 #include "FFlBUSH.H"
-#include "FFlBEAM2.H"
-#include "FFlBEAM3.H"
-#include "FFlTRI3.H"
-#include "FFlTRI6.H"
-#include "FFlQUAD4.H"
-#include "FFlQUAD8.H"
-#include "FFlTET4.H"
-#include "FFlTET10.H"
-#include "FFlWEDG6.H"
-#include "FFlWEDG15.H"
-#include "FFlHEX8.H"
-#include "FFlHEX20.H"
-#include "FFlCLOAD.H"
-#include "FFlPLOAD.H"
+#include "FFlBeams.H"
+#include "FFlShells.H"
+#include "FFlSolids.H"
+#include "FFlLoads.H"
 #include "FFlPMASS.H"
 #include "FFlPSPRING.H"
 #include "FFlPBUSHCOEFF.H"
@@ -47,9 +37,6 @@
 #include "FFlPWAVGM.H"
 #include "FFlPRBAR.H"
 #include "FFlPMAT.H"
-#include "FFlPMAT2D.H"
-#include "FFlPMAT3D.H"
-#include "FFlPMATSHELL.H"
 #include "FFlPCOORDSYS.H"
 #ifdef FT_USE_STRAINCOAT
 #include "FFlSTRCoat.H"
