--- FFlLib/FFlFEParts/FFlWEDG15.H
@@ -1,50 +1,0 @@
-// SPDX-FileCopyrightText: 2023 SAP SE
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-// This file is part of FEDEM - https://openfedem.org
-////////////////////////////////////////////////////////////////////////////////
-
-#ifndef FFL_WEDG15_H
-#define FFL_WEDG15_H
-
-#include "FFlLib/FFlElementBase.H"
-#include "FFlLib/FFlFEElementTopSpec.H"
-#include "FFlLib/FFlFEAttributeSpec.H"
-#include "FFlLib/FFlTypeInfoSpec.H"
-#ifdef FT_USE_MEMPOOL
-#include "FFaLib/FFaPatterns/FFaMemPool.H"
-#endif
-#include "FFaLib/FFaPatterns/FFaSingelton.H"
-
-
-class FFlWEDG15 : public FFlElementBase
-{
-public:
-  FFlWEDG15(int ID) : FFlElementBase(ID) {}
-  virtual ~FFlWEDG15() {}
-
-  static void init();
-
-  FFL_FE_ELEMENT_FACTORY_INIT(FFlWEDG15);
-  FFL_FE_ELEMENT_TOP_SPEC(FFlWEDG15);
-  FFL_FE_ATTRIBUTE_SPEC(FFlWEDG15);
-  FFL_TYPE_INFO(FFlWEDG15);
-
-  virtual bool getFaceNormals(std::vector<FaVec3>& normals, short int face = 1,
-			      bool switchNormal = false) const;
-
-  virtual bool getVolumeAndInertia(double& volume, FaVec3& cog,
-				   FFaTensor3& inertia) const;
-
-#ifdef FT_USE_MEMPOOL
-  FFA_MAKE_MEMPOOL;
-#endif
-};
-
-
-typedef FFaSingelton<FFlTypeInfoSpec,FFlWEDG15>     FFlWEDG15TypeInfoSpec;
-typedef FFaSingelton<FFlFEAttributeSpec,FFlWEDG15>  FFlWEDG15AttributeSpec;
-typedef FFaSingelton<FFlFEElementTopSpec,FFlWEDG15> FFlWEDG15ElementTopSpec;
-
-#endif
