+++ FFlLib/FFlNastranProcessor.C
@@ -19,9 +19,6 @@
 #include "FFlLib/FFlFEParts/FFlPCOMP.H"
 #include "FFlLib/FFlFEParts/FFlPMASS.H"
 #include "FFlLib/FFlFEParts/FFlPMAT.H"
-#include "FFlLib/FFlFEParts/FFlPMAT2D.H"
-#include "FFlLib/FFlFEParts/FFlPMAT3D.H"
-#include "FFlLib/FFlFEParts/FFlPMATSHELL.H"
 #include "FFlLib/FFlFEParts/FFlPNSM.H"
 #include "FFlLib/FFlFEParts/FFlPRGD.H"
 #include "FFlLib/FFlFEParts/FFlPRBAR.H"
